Set up and start the queue for run_single_step in python
    start and montor job queue using job_queue_manager in python

    run job in a separate thread monitoring themselves

    execute job_queue_node  callbacks when job is not longer running

    Keep callbacks driver and jobs in python

bump libecl

+++ lib/job_queue/local_driver.cpp
@@ -90,7 +90,8 @@ void local_driver_free_job( void * __job ) {
 
 void local_driver_kill_job( void * __driver , void * __job) {
   local_job_type    * job  = local_job_safe_cast( __job );
-  kill( job->child_process , SIGTERM );
+  if (job->child_process > 0)
+    kill( job->child_process , SIGTERM );
 }
